py/gf_ml/gf_ml_adapters/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
google-auth